+++ Tests/TuistKitTests/Generator/Mocks/MockManifestLoader.swift
@@ -22,6 +22,14 @@ final class MockGraphManifestLoader: GraphManifestLoading {
     var loadTuistConfigCount: UInt = 0
     var loadTuistConfigStub: ((AbsolutePath) throws -> ProjectDescription.TuistConfig)?
 
+    var loadGalaxyCount: UInt = 0
+    var loadGalaxyStub: ((AbsolutePath) throws -> ProjectDescription.Galaxy)?
+
+    func loadGalaxy(at path: AbsolutePath) throws -> ProjectDescription.Galaxy {
+        loadGalaxyCount += 1
+        return try loadGalaxyStub?(path) ?? ProjectDescription.Galaxy.test()
+    }
+
     func loadProject(at path: AbsolutePath) throws -> ProjectDescription.Project {
         return try loadProjectStub?(path) ?? ProjectDescription.Project.test()
     }